created helper function file and fixed /update
